// freelancer-client/src/features/sellers/interfaces/seller.interface.ts
export interface IPersonalInfoData {
  [key: string]: string;
  fullName: string;
  // profilePicture: string;
  description: string;
  responseTime: string;
  oneliner: string;
}

export interface IPersonalInfoProps {
  personalInfo: IPersonalInfoData;
  setPersonalInfo: Dispatch<SetStateAction<IPersonalInfoData>>;
  personalInfoErrors: IPersonalInfoData[];
}